+++ client/src/containers/index.ts
@@ -1,5 +1,5 @@
 export {default as ScaledOrders} from './ScaledOrders/ScaledOrders';
-export {default as MarketOrderContainer} from './MarketOrder/MarketOrder';
+export {default as MarketOrderContainer} from './MarketOrder/MarketOrderContainer';
 export {default as TrailingLimitOrder} from './TrailingLimitOrder/TrailingLimitOrderContainer';
 export {default as TickerPricesContainer} from './TickerPrices/TickerPricesContainer';
 export {default as CrossOrderContainer} from './CrossOrder/CrossOrderContainer';
